+++ web-ts/src/actions/ActionEvent.tsx
@@ -1,5 +1,5 @@
 enum ActionEvent {
-  GET_MENU_LIST = 'GET_MENU_LIST',
+  RECEIVE_MENU_LIST = 'RECEIVE_MENU_LIST',
   GET_USER_INFO = 'GET_USER_INFO',
 }
